Homework 4/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
gunicorn==22.0.0

# Azure SDKs
azure-cosmos==4.7.0
azure-storage-blob==12.20.0
azure-servicebus==7.12.2
azure-ai-textanalytics==5.3.0
azure-core==1.30.2
